reversed the UX
Previously one would open the search input, make the search, and the
pages would open. This made sense when using console.lua, because the
console logs covered up the page making it ugly to look at and read.
However, with the new input, it now makes more sense to have the keybind
open the page, and then use a dynamic keybind to open the search box.